21 release clusttmb 010

+++ tests/testthat/test-unit-utils-fittmb.R
@@ -8,6 +8,8 @@ context("unit test utils-fit.tmb")
 
 
 test_that("run.model = FALSE", {
+  skip_on_cran()
+  skip_if_not(run_model)
   clustTMB.mod <- fit.tmb(args,
     control = list(run.model = FALSE)
   )
@@ -26,6 +28,8 @@ test_that("run.model = FALSE", {
 })
 
 test_that("run.model = TRUE, sdr.report = FALSE", {
+  skip_on_cran()
+  skip_if_not(run_model)
   clustTMB.mod <- fit.tmb(args,
     control = list(
       run.model = TRUE,
@@ -56,6 +60,8 @@ test_that("run.model = TRUE, sdr.report = FALSE", {
 })
 
 test_that("run.model = TRUE, sdr.report = TRUE", {
+  skip_on_cran()
+  skip_if_not(run_model)
   clustTMB.mod <- fit.tmb(args,
     control = list(
       run.model = TRUE,
@@ -86,6 +92,8 @@ test_that("run.model = TRUE, sdr.report = TRUE", {
   expect_equal(sdr, clustTMB.mod$sdr)
 })
 
-dyn.unload(dynlib(testthat::test_path("ref", "simple")))
-file.remove(dynlib(testthat::test_path("ref", "simple")))
-file.remove(testthat::test_path("ref", "simple.o"))
+if(run_model){
+  dyn.unload(dynlib(testthat::test_path("ref", "simple")))
+  file.remove(dynlib(testthat::test_path("ref", "simple")))
+  file.remove(testthat::test_path("ref", "simple.o"))
+}
